Create sumOfMultiplicationvalues.c
Program Description

This C program is designed to calculate the sum of the first n natural numbers using a for loop.

The program begins by asking the user to input a positive integer n, which represents how many natural numbers will be added together.

It then initializes a variable sum to zero.

A for loop is used to run from 1 to n. In every iteration of the loop, the current number (i) is added to the variable sum.

Once the loop ends, the variable sum holds the total of all natural numbers from 1 up to n.

Finally, the program prints the calculated sum as the output.

This program demonstrates:

Basic input/output operations using printf and scanf.

Use of a counter-controlled loop (for loop) to perform repetitive addition.

Concept of accumulation, where a single variable (sum) is updated continuously to store the result.

For example:

If the user enters n = 5, the program adds 1 + 2 + 3 + 4 + 5 = 15 and prints 15.